% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Descriptive.R
\name{print.diagonal}
\alias{print.diagonal}
\title{Print the diagonal of the affinity matrix}
\usage{
\method{print}{diagonal}(
  res,
  path = getwd(),
  name = "diagonal_affinity_matrix",
  labels = paste0("Trait ", 1:K),
  pr = 0.05
)
}
\arguments{
\item{res}{A list corresponding to the output of \code{\link{estimate.affinity.matrix}}, \code{\link{estimate.affintiy.matrix.lowrank}} or \code{\link{estimate.affinity.matrix.unipartite}}.}

\item{path}{A string indicating the path where to save the txt file. Defaults to current path.}

\item{name}{A string indicating the name of the txt file. Defaults to \code{"diagonal_affinity_matrix"}.}

\item{labels}{A vector of strings indicating the names of the matching variables. Defaults to \code{"Trait k"} for every \code{k} matching variable.}

\item{pr}{A probability indicating the two-tailed significance level required for an estimated parameter to be printed in boldface. Defaults to 0.05 and can be set to 0 to avoid printing any estimate in boldface.}
}
\value{
The function saves a txt file in tex style in \code{path}.
}
\description{
This function prints the estimated diagonal elements of the affinity matrix in tex style and stores them in a txt file.
}
\details{
For the sake of clarity, we refer to "men" as the observations on one side of the market and to "women" as the observations on the other side.
}
