% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Descriptive.R
\name{show.saliency}
\alias{show.saliency}
\title{Print summary of saliency analysis}
\usage{
show.saliency(
  res,
  path = getwd(),
  name_x = "saliency_analysis_x",
  name_y = "saliency_analysis_y",
  ncol_x = Kx,
  ncol_y = Ky,
  labels_x = paste0("Trait ", 1:Kx),
  labels_y = paste0("Trait ", 1:Ky),
  pr = 0.05
)
}
\arguments{
\item{res}{A list corresponding to the output of \code{\link{estimate.affinity.matrix}}, \code{\link{estimate.affintiy.matrix.lowrank}} or \code{\link{estimate.affinity.matrix.unipartite}}.}

\item{path}{A string indicating the path where to save the txt file. Defaults to current path.}

\item{name_x}{A string indicating the name of the txt file storing the results for men. Defaults to \code{"saliency_analysis_x"}.}

\item{name_y}{A string indicating the name of the txt file storing the results for women. Defaults to \code{"saliency_analysis_y"}.}

\item{ncol_x}{An integer indicating the number of singular vector to print for men. The function prints the first \code{ncol_x} singular vectors. Defaults to \code{ncol(U)}.}

\item{ncol_y}{An integer indicating the number of singular vector to print for women. The function prints the first \code{ncol_y} singular vectors. Defaults to \code{ncol(V)}.}

\item{labels_x}{A vector of strings indicating the names of men's matching variables. Defaults to \code{"Trait k"} for every \code{k} matching variable.}

\item{labels_y}{A vector of strings indicating the names of women's matching variables. Defaults to \code{"Trait k"} for every \code{k} matching variable.}

\item{pr}{A probability indicating the two-tailed significance level required for an estimated parameter to be printed in boldface. Defaults to 0.05 and can be set to 0 to avoid printing any estimate in boldface.}
}
\value{
The function saves a txt file in tex style in \code{path}.
}
\description{
This function prints the results from the saliency analysis in tex style and stores them in a txt file.
}
\details{
For the sake of clarity, we refer to "men" as the observations on one side of the market and to "women" as the observations on the other side.
}
