% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Descriptive.R
\name{show.test}
\alias{show.test}
\title{Print summaries of rank tests}
\usage{
show.test(res, path = getwd(), name = "rank_tests", pr = 0.05, n_tests = K - 1)
}
\arguments{
\item{res}{A list corresponding to the output of \code{\link{estimate.affinity.matrix}}, \code{\link{estimate.affinity.matrix.lowrank}} or \code{\link{estimate.affinity.matrix.unipartite}}.}

\item{path}{A string indicating the path where to save the txt file. Defaults to current path.}

\item{name}{A string indicating the name of the txt file. Defaults to \code{"rank_tests"}.}

\item{pr}{A probability indicating the significance level required to pass a rank test. Defaults to 0.05.}

\item{n_tests}{An integer indicating the number of tests to show. The function prints the first \code{n_tests} rank tests. Defaults to \code{min(nrow(Y),nrow(X))-1}.}
}
\value{
The function saves a txt file in tex style in \code{path}.
}
\description{
This function prints the summaries of the first \code{n_tests} rank tests in tex style and stores them in a txt file.
}
\details{
For the sake of clarity, we refer to "men" as the observations on one side of the market and to "women" as the observations on the other side.
}
